Acc g-force conversion and EOL tool testing sensors, output to logs
For some behaviours under"basic_behaviours", I converted and printed
Acceleromter reading to mg-force and analogue readings to % of max
value.
For the EOL Tool, the sensors values will not be displayed and tester
need to check if the values make sense. The results will be saved to a
log file in the qa_log folder.