Test Bed initial experiment set up
Tuned parameters to enable expected behaviours
Added Idling preferred feature; it will do idle when the reward is low
in all region
The idling reward is smoothed by taking average over a window of time.
Also, the it has to be non-idle for period time after breaking out of
idle mode
# of exemplars threshold will increase every split
If split quality as measured by within-group variance is low, it won't
split. The threshold inherited from parents times a decay factor which
approaches one.
Added a function that save figures to files